Fix but with with checked="checked"
http://stackoverflow.com/questions/6637723/jquery-uniform-checkbox-does-not-uncheck

Query 1.6.4 changed the meaning of the attr()
function, which is used by jquery.uniform.js.
attr() now returns the state that the attribute
WAS in when the page loaded - not what the
state of the attribute is NOW. There is a
replacement function called prop(), which does
the job that attr()
used to do.

To fix the plugin, replace each occurrence
of attr("checked") with prop("checked").
Also, change attr("disabled") to
prop("disabled").

That should fix your project up. It worked for
me ;-)

See also:
https://github.com/pixelmatrix/uniform/pull/167,
which attempts to address the issue with better
backwards compatibility.